+++ samples/EntityFrameworkCore.GenericRepository.Samples.WebApi/Startup.cs
@@ -4,6 +4,7 @@ using EntityFrameworkCore.GenericRepository.Abstractions;
 using EntityFrameworkCore.GenericRepository.Samples.Shared;
 using Microsoft.AspNetCore.Builder;
 using Microsoft.AspNetCore.Hosting;
+using Microsoft.EntityFrameworkCore;
 using Microsoft.Extensions.DependencyInjection;
 using Newtonsoft.Json;
 using Newtonsoft.Json.Serialization;
@@ -20,8 +21,10 @@ namespace EntityFrameworkCore.GenericRepository.Samples.WebApi
                     options.SerializerSettings.ContractResolver = new CamelCasePropertyNamesContractResolver();
                     options.SerializerSettings.NullValueHandling = NullValueHandling.Ignore;
                 });
+
+            services.AddDbContext<DbContext, SampleDbContext>(ServiceLifetime.Transient);
             
-            services.AddTransient<GenericRepositoryContext, SampleDbContext>();
+            services.AddTransient<DbContext, SampleDbContext>();
             services.AddTransient(typeof(IEntityRepository<,>), typeof(EntityRepository<,>));
 
             services.AddAutoMapper(config => config.AddCollectionMappers(), typeof(Startup).Assembly);