+++ src/EntityFrameworkCore.GenericRepository/GenericRepositoryContext.cs
@@ -1,9 +1,0 @@
-﻿using Microsoft.EntityFrameworkCore;
-
-namespace EntityFrameworkCore.GenericRepository
-{
-    public abstract class GenericRepositoryContext : DbContext
-    {
-        protected abstract override void OnConfiguring(DbContextOptionsBuilder optionsBuilder);
-    }
-}