+++ test/EntityFrameworkCore.GenericRepository.TestInfrastracture/AddressDto.cs
@@ -1,0 +1,12 @@
+namespace EntityFrameworkCore.GenericRepository.TestInfrastracture
+{
+    public class AddressDto
+    {
+        public AddressDto(string city)
+        {
+            this.City = city;
+        }
+
+        public string City { get; }
+    }
+}